requirements.txt
streamlit
langchain-core
python-dotenv
langchain-openai
openai
langchain-anthropic
langchain-groq
shutil
pathlib